refactor: Extract domain models to separate files following SRP
