+++ internal/domain/morningcall.go
@@ -1,0 +1,15 @@
+package domain
+
+import (
+	"time"
+)
+
+// アラーム情報
+type MorningCall struct {
+	ID         MorningCallID
+	SenderID   string
+	ReceiverID string
+	Time       time.Time
+	Message    string
+	Status     MorningCallStatus
+}
