+++ internal/domain/morningcall_status.go
@@ -1,0 +1,10 @@
+package domain
+
+type MorningCallStatus string
+
+const (
+	MorningCallStatusScheduled MorningCallStatus = "scheduled"
+	MorningCallStatusDeleted   MorningCallStatus = "deleted"
+	MorningCallStatusCompleted MorningCallStatus = "completed"
+	MorningCallStatusFailed    MorningCallStatus = "failed"
+)
