+++ internal/domain/morningcall_validation.go
@@ -1,0 +1,3 @@
+package domain
+
+// MorningCall validation methods will be added here
