+++ internal/domain/related_user_status.go
@@ -1,0 +1,10 @@
+package domain
+
+type RelatedUserStatus string
+
+const (
+	RelatedUserStatusApproved RelatedUserStatus = "approved"
+	RelatedUserStatusPending  RelatedUserStatus = "pending"
+	RelatedUserStatusRejected RelatedUserStatus = "rejected"
+	RelatedUserStatusBlocked  RelatedUserStatus = "blocked"
+)
